added lower-level tools for building bson.
I discovered a problem related to using NSDictionary instances to
build bson queries. NSDictionary keys are unordered, so there is
no way to control the ordering of elements in a bson object
created from a dictionary. This is ordinarily not a problem,
but commands run with the runCommand:inDatabase: method must
have their elements specified in the correct order.